Restricted Users for MainUser
Previously, the parent of a restricted user had to be the system user
(in the creation method). On HSUM, we want this to be the main user. The
canHaveRestrictedProfile() method already took this into account, but
not the actual creation method. We fix this here.

Test: on a device without feature android.hardware.telephony, do the following:
make GtsGmsCoreMultiUserTestApp  &&    atest GtsGmscoreHostTestCases:com.google.android.gts.multiuser.RestrictedProfileHostTest
Bug: 414839808
Flag: EXEMPT bugfix
Change-Id: Id69ead7224ee901a18a5153f7627a7cd0b3f45ce