Get bubble bar distance from launcher
Update Bubbles AIDL interface to pass the distance between the
bubble bar and the bottom of the screen rather than the top
coordinate of the bubble bar. When the screen height changes,
launcher has a short delay and calculates the bubble bar top based
on the previous screen height. This change moves the offsetting to
the shell side.

Bug: 392893178
Flag: com.android.wm.shell.enable_bubble_bar
Test: manual
       - send some bubbles
       - launch app
       - expand
       - swipe to home
       - fold and unfold

Change-Id: Ibf6d141ac72b354c275a4b58c4618db08460e788