Add batch update bindings API.
This will allow applications to reduce the number of calls to AMS when
multiple bindings need to be updated at once (e.g. for multi process
applications).  It will allow for a reduction in the number of
OomAdjuster updates and also allow for a small number of bindings
overall since bindings can be updated instead of having to use multiple
bindings as priorities change to services.

Changes include:
- Adding enable_update_service_bindings flag
- Adding Context.UpdateBindingParams
- Adding Context.updateServiceBindings via ActiveServices/ActivityManagerService
- Adding internal BindUpdateInfo interface
- Splitting up ActiveServices.unbindServiceLocked
- Adding ConnectionRecordInternal.updateFlags

Bug: 424003130
Flag: android.content.flags.enable_update_service_bindings
Test: atest android.app.cts.service.ServiceTest
Change-Id: If956709f5556e770a79345c0762c21181858e1ac