Delay CompatChanges evaluation for CT enabled default value
The default value of CT verification was evaluated as a static final
field. The early evaluation caused issues in the
`CtsNetSecConfigCertificateTransparencyDefaultSdk36TestCases` test, as
it appears the app compat environment was not initialized properly yet.

To fix the problem,
`DEFAULT_CERTIFICATE_TRANSPARENCY_VERIFICATION_REQUIRED` was replaced
with a function. The function computes the value on demand, and it is
now called when the `CompatChanges` initialization is finished.

Since the original implementation of this CL causes some memory
regressions (b/415229614), in this version we lazily evaluate the
default enablement flags and cache the result.

Bug: 407952621
Test: atest CtsNetSecConfigCertificateTransparencyDefaultSdk36TestCases
Flag: com.android.org.conscrypt.net.flags.certificate_transparency_default_enabled
Change-Id: I790f4a9e71b9b8afd14bfcfd39ed3d273c1872da