SFPS indicator becomes visibile after position is ready
Don't set the visibility of the SFPS indicator to VISIBLE
until after the window parameters of the view are set.
This prevents a flash of the view being in the wrong location.

We also update repeatOnLifecycle(STARTED) to CREATED
so that we are able to collect the window params before
the view is visible.

Flag: EXEMPT bugfix
Test: enroll sideFP, show bouncer on lockscreen,
observe correct position of SFPS indicator (no flash
of it being in the wrong place)
Fixes: 398820784

Change-Id: Ia27905072b3bfcdd86e18a8a3c86e30692778b53