Implement BundleHeader guts in Compose
This is a conservative version without changing any guts APIs at this
point. See BundleHeaderGutsContent class documentation for additional
context. I may attempt to remove the View in a follow up CL but I
wanted to not complicate this CL which focuses on adding the Compose UI
and enabling it.

Most functionality is in TODOs for follow-up CLs so we can iterate fast
on this base setup.

Test: Manual test and tests added
Bug: 409748420
Flag: com.android.systemui.notification_bundle_ui
Change-Id: Ic6e268ea854ca265f6cb0db49bdd297514baefa1