Update surface position on removing leash for a closing window
An animating WindowState will set surface position on animation
leash and its surface will be set to (0,0), when the animation
is finished, the animation leash will be removed and the actual
surface position needs to restore to WindowState.

Because isGoneForLayout returns true if the activity is invisible
requested, then the window may miss to restore surface position.

Bug: 381968939
Flag: EXEMPT bugfix
Test: Open AppInfo. Either click Disable/Uninstall/ForceStop.
      And then click the list item immediately (e.g. Wi-Fi data).
      The dialog window should not flash on top (zero position).
Change-Id: I1a580a0d37c56fc30b3995cbbe6bd4b4d3ca6735