[SB][Call] Share click listener code between call chip and notif chip.
Now that the call chip and the notification chips have the same click
behavior (show the associated notification), we should have them share
code.

This is mostly a no-op change, except that it makes the call chip click
listener correctly be `HideHeadsUpNotification` if the call notification
is already showing.

Bug: 414830065
Flag: android.app.ui_rich_ongoing
Test: atest NotifChipsViewModelTest CallChipViewModelTest
Test: Verify a11y content description of call chip changes based on HUN
status
Test: Verify chip clicks are still in log buffers
Change-Id: I360403a67a3caa8cc5278716f6b805371d0d7aae

Change-Id: I3e3376d6faae85fc26c443758e4e6b7567d127e0