Use WindowRelayoutResult for all WM->Client layout
The interface between client and WM has multiple methods
passing around the same set of information. It turns out that
WindowRelayoutResult encapsulates this pretty well and is
parcelable which makes it work well as a binder `out` parameter.

Take the existing APIs which are passing layout information and
have them all use this one.

This includes pulling the SurfaceControl out of WindowRelayoutResult
since it is only used for relayout -- in other words, common stuff
stays in WindowRelayoutResult while each API's unique parameters
don't.

This reverts commit 39aef49c9e4c78deca40cef75a038f463094a19f.

Reason for revert: To re-apply after fixing robolectric shadow.

Bug: 385976595
Bug: 417782119
Change-Id: I842b98716369e7074f4ec8c7dc516d0644465328