Refactor: rename InputDevice specific vibration provider methods
There used to be 2 overloaded methods with the same name: getVibration.
This CL renames one of them to getVibrationForInputDevice. This
simplifies reading usages of this method, and also aligns with how the
getVibrationAttributes and getVibrationAttributesForInputDevice methods
of the same class are named.
Similarly, the InputDevice specific method in
HapticFeedbackCustomization has been renamed to follow the same pattern.

Bug: 397602072
Bug: 408393305
Test: atest HapticFeedbackVibrationProviderTest
Flag: android.os.vibrator.haptic_feedback_with_custom_usage
Change-Id: I46ed5e3e25692b153617c107006333cc42bad209