Always return the remaining velocity after a fling
This CL slightly changes the logic in SwipeAnimation so that we always
return the velocity remaining after animating to the target offset, even
if the animation was done using a spring. This allows to still have a
slight overscroll when the spring is animating from a large offset, and
there is no reason not to have it as it arguably feels better.

Bug: 417444347
Test: Manual, noticed the slight overscroll even when not decaying
Flag: com.android.systemui.scene_container
Change-Id: I33ff9e134ec97dadb4f95d86709da9224c67065d