[PiP on CD] Activate desk if expanding PiP at Home on Desktop-first
displays in multi-desks case.

When expanding to PiP at Home on Desktop-first displays, we would expect
PiP to expand to freeform windowing mode, just like launching any other
task. However, at the time of expanding PiP, if it is the only task
running, there is actually no Desks active yet. In this case, we also
need to add desk activation changes, in addition to reparenting the PiP
task to the root desk.

Note that this change also relies on the
enable_overview_on_connected_displays flag to update the Taskbar with
the task icon.

Bug: 417522949
Test: atest WMShellUnitTests:com.android.wm.shell.pip2
Test: atest DesktopPipTransitionControllerTest
Test: Manual - expand PiP at Home on Desktop-first display, verify PiP
icon is added to Taskbar and reparented to root desk
Flag: com.android.window.flags.enable_connected_displays_pip

Change-Id: Ibb025c17e1a787171800295a188db9ba6c1a80b3