Make the default system session non-releasable
The default session is what we show to unprivileged apps as the current
"system output". For example, the built-in speaker, BT, HDMI, etc. By
definition, it's what we show to apps with no privilege, and hence the
release type is always going to be "non-releasable".

Also use RELEASE_UNSUPPORTED if the personal audio sharing flag is
disabled.

Test: Manually using the AndroidX sample routing app.
Bug: 385672684
Flag: com.android.media.flags.enable_output_switcher_personal_audio_sharing
Change-Id: I454082dd44a78d38f2836e97861a13174b5a3fbe