3.1 Extract logic from startIntentWithAnimation into a helper.
This is to isolate the branching from the entry point. Follow up
will add a new entry point using the new APIs, and the branching
will be inside this helper.

Trying to make each of the CLs as small as possible to keep them
digestible and low risk. For the refactor plan see
go/animlib-shell-refactor-plan.

Bug: 397180418
Flag: EXEMPT simple refactor
Test: atest ActivityTransitionAnimatorTest + manual
Change-Id: If10df0fbd43da3c58711ff329e4bc8d378f01486