Report app handle bounds to status bar
Adds an AppHandleIdentifier to CaptionState and has AppHandleNotifier
listen to caption state flow.

Also updates caption state even when task is not focused since app
handle position should be updated for both tasks in split, and in split
on of the tasks is always not focused.

Bug: 412444130
Test: check bounds for split and fullscreen
Flag: com.android.window.flags.enable_app_handle_position_reporting
Change-Id: I96c23602c16000b77dd70c6d48d3829568e01146