Improve preview icons animation
The original spec used a spring for the icons. Unfortunately, that's not
easy to implement with STL as transition is driven by a 0 -> 1 progress.
Jordan suggested that we may write a CustomTransformation that triggers
a spring. We can attempt to do this in a future improvement of the
animation.

I tried using a curve that is translated from a spring "Expressive
default spatial" found here:
https://carbon.googleplex.com/google-material-3/pages/motion/specs
Unfortunately, there seems to be a bug in STL where the overshoot of the
curve is clamped to 1.0f, therefore it appears as if the icon stops
moving when it overshoots.

I talked to jdepriest@ to find a variation that uses our old curves.
It is a bit less bouncy as it doesn't overshoot but we agreed this
version looks great for now. We can iterate later on.

Bug: b/417651636
Test: Verified animation in GalleryApp and video signed off by
      Motion Designer jdepriest@google.com
Flag: com.android.systemui.notification_bundle_ui
Change-Id: Iea1d6e31e7aae6f5f3262323b858524ede261bc5