Notify non-top visible activities to start if no paused activity
Given the path when launching an activity:
realStartActivityLocked
 ensureVisibilityAndConfig(deferResume=true)
  ensureActivitiesVisible(notifyClients=false)
   Skip makeActiveIfNeeded by notifyClients
 Skip ensureConfigAndVisibilityAfterUpdate by deferResume

When launching a translucent activity on an existing non-fullscreen
task with moving the task to front, because there is no activity
needs to pause, there won't be invocation of ensureActivitiesVisible
with notifyClients=true. Then the non-top activity of the task may
stay at visible-requested but the client doesn't get lifecyle change
of onStart to make the activity visible and draw.

So manually invoke makeActiveIfNeeded for the activities of the task.

Bug: 418943399
Flag: EXEMPT bugfix
Test: atest ActivityTaskSupervisorTests# \
      testNonTopVisibleActivitiesActiveWhenLaunchingTranslucent
Change-Id: I926cb6bec3475bb9830c1316b1343c0163dcadd5