chore(a11yInputFilter): dump received and sent events when exceptions in AccessibilityInputFilter
Cache the events when the AccessibilityInputFilter onInputEvent and
sendInputEvent methods are called, then dump the events when sendInputEvent gets an
expection (implies not succeed), so that we can have more data to investigate
b/414179573.

The dump only works when the original DEBUG flag true (implies
Log.debuggable) or the build is debuggable, so by default it's an no-op.

Bug: 414179573
Flag: EXEMPT only add logging for bug investigation
Test: manually check logcat & local bug report
Change-Id: I7ecdc94e88a68e474e02b6ef5249254290ec11d3