Fix broken CUJ tacking for CUJ_LPP_ASSIST_INVOCATION_EFFECT
This fixes the issue that the wrong Choreographer was used for the
vsync id. Since the animator runs in SystemUI, the vsyncId obtained from
the Choreographer from SystemUI needs to be used.

Secondly, for correct CUJ tracking, the handler from the Thread that
runs the animators needs to be passed to InteractionJankMonitor#begin.

Bug: 418136893
Test: TopLevelWindowEffectsTest
Test: Manual, i.e. verified that CUJ is correctly showing up in perfetto
      trace
Flag: com.android.systemui.shared.enable_lpp_assist_invocation_effect
Change-Id: I125d7b7e6236b0858c52490ffc361b2f1695767b