Reschedule bg stop if profile has alarm/audio
We avoid stopping a background user if it is playing audio or has an
upcoming alarm. But, currently, if its profile has audio/alarm, we still
may stop the parent, which will automatically stop the profile too.
We now make sure that if the profile shouldn't be stopped, we also don't
stop the parent (and therefore don't wind up indirectly stopping the
profile).

Bug: 401340391
Bug: 330351042
Test: UserControllerTest
Flag: android.multiuser.schedule_stop_of_background_user
Change-Id: I2341891ff193cbbb86cb2b0371e3a7b858c020bb