Fix VibrationThread handling of IVibrator errors
The vibration playback is ignoring the HAL result for vibrate requests,
causing the vibration to continue even after one of the steps
is unsupported or failed.

This also causes the vibration request to be reported as successful in
the vibrator service dumpsys, which makes debugging difficult when the
user reports no vibration was actually performed by the device.

This fix introduces a new status code for HAL errors and updates the
VibrationThread to stop the vibration steps once a step reports
unsupported or failed status.

Bug: 419572960
Flag: android.os.vibrator.vibration_thread_handling_hal_failure
Test: VibrationThreadTest
Change-Id: Ib6a8ea6041185212d07204b2ee24e0d592c51083