Log notification updates when there's visually changed text.
Prior to this change we only log if the notification is alerting (or would be alerting but is muted from cooldown), or if one of a particular set of fields has changed. New messages to an ongoing conversation do not change any of these fields, so they were only logged as updates if they were alerting, and not if it was just new text coming in.

Bug: 420989666
Flag: com.android.server.notification.log_visually_changed_updates
Test: manual via atom tester
Change-Id: Ib099c5684fcfdb1c9835d7ddcab75419bc759511