Dont record adjacent display if non-internal
This fixes the bug where the FLAG_DEFAULT_GROUP_ADJACENT is set for
virtual displays on cuttle fish device. This is because the desktop mode
is not enabled on these devices, which forces the content mode for these
displays to be set to non-desktop and makes it eligible for the flag. We
are now explicitly adding the check of internal and external displays
before setting the flag

Bug: 414444200
Flag: EXEMPT Trivial Change
Test: Manual
Change-Id: I3b2afb75d3bdbc17e197a60bdfbb1604787f097f