Return new error codes in VerifyCredentialResponse
If the SoftwareRateLimiter reports that the credential is too short or
was already tried, report this in the VerifyCredentialResponse so that
upper layers such as Keyguard will have access to this information.

Similarly, if Weaver reports INCORRECT_KEY and a timeout of zero, report
this as VerifyCredentialResponse#RESPONSE_CRED_INCORRECT so that
LockSettingsService#reportResultToSoftwareRateLimiter() and upper layers
such as Keyguard will have access to this information.

Note that although these new error codes will now be present in the
VerifyCredentialResponse, ultimately this does not change any other
behavior yet, since no non-test code treats RESPONSE_OTHER_ERROR and
unknown errors differently yet.  (Any such user would have come up in
the rename of RESPONSE_ERROR to RESPONSE_OTHER_ERROR.)  This CL just
makes it possible to handle different errors differently in later CLs.

Test: atest FrameworksServicesTests:com.android.server.locksettings
Bug: 395976735
Flag: android.security.software_ratelimiter
Change-Id: I5fdc2fc54b841dc5889c2f92b1f9f3220ad70d98